moved code into price-per-unit-helper
