+++ config/app_config.php
@@ -148,7 +148,8 @@ return [
         'timeHelper' => new App\View\Helper\MyTimeHelper(new Cake\View\View()),
         'numberHelper' => new App\View\Helper\MyNumberHelper(new Cake\View\View()),
         'slugHelper' => new App\View\Helper\SlugHelper(new Cake\View\View()),
-        'timebasedCurrencyHelper' => new App\View\Helper\TimebasedCurrencyHelper(new Cake\View\View())
+        'timebasedCurrencyHelper' => new App\View\Helper\TimebasedCurrencyHelper(new Cake\View\View()),
+        'pricePerUnitHelper' => new App\View\Helper\PricePerUnitHelper(new Cake\View\View())
     ],
     'DateFormat' => [
         'Database' => 'yyyy-MM-dd',
