+++ src/View/Helper/MyHtmlHelper.php
@@ -37,59 +37,6 @@ class MyHtmlHelper extends HtmlHelper
         parent::__construct($View, $config);
     }
     
-    public function getQuantityInUnitsStringForAttributes($attributeName, $attributeCanBeUsedAsUnit, $unitPricePerUnitEnabled, $unitQuantityInUnits, $unitName, $amount=1)
-    {
-        $quantityInUnitsString = $this->getQuantityInUnits($unitPricePerUnitEnabled, $unitQuantityInUnits, $unitName, $amount);
-        if ($quantityInUnitsString != '') {
-            $unitName = $attributeName . ', ' . $quantityInUnitsString;
-        }
-        if ($attributeCanBeUsedAsUnit) {
-            $unitName = $quantityInUnitsString;
-        }
-        if ($unitName == '') {
-            $unitName = $attributeName;
-        }
-        return $unitName;
-    }
-    
-    public function getQuantityInUnitsWithWrapper($quantityInUnitsEnabled, $quantityInUnits, $unitName)
-    {
-        return '<span class="quantity-in-units">' . $this->getQuantityInUnits($quantityInUnitsEnabled, $quantityInUnits, $unitName) . '</span>';
-    }
-    
-    public function getQuantityInUnits($quantityInUnitsEnabled, $quantityInUnits, $unitName, $amount=1)
-    {
-        $result = '';
-        if ($quantityInUnitsEnabled && $quantityInUnits > 0) {
-            if ($amount > 1) {
-                $result  = 'je ' . $result;
-            }
-            $result .= 'ca. ' . $this->formatAsDecimal($quantityInUnits, 2, true) . ' ' . $unitName;
-        }
-        return $result;
-    }
-
-    public function getPricePerUnit($priceInclPerUnit, $quantityInUnits, $amount)
-    {
-        return '<div class="price">' . $this->formatAsEuro($priceInclPerUnit * $quantityInUnits / $amount) . '</div> <div class="price-asterisk">*</div>';
-    }
-    
-    public function getPricePerUnitInfoText($priceInclPerUnit, $unitName, $unitAmount)
-    {
-        $infoText = '<div class="line">';
-            $infoText .= '<span class="additional-price-info">';
-            $infoText .= ' * Basis-Preis: ' . $this->getPricePerUnitBaseInfo($priceInclPerUnit, $unitName, $unitAmount);
-            $infoText .= ', Preis wird evtl. noch angepasst.';
-            $infoText .= '</span>';
-        $infoText .= '</div>';
-        return $infoText;
-    }
-    
-    public function getPricePerUnitBaseInfo($priceInclPerUnit, $unitName, $unitAmount)
-    {
-        return $this->formatAsEuro($priceInclPerUnit) . ' / ' . ($unitAmount > 1 ? $this->formatAsDecimal($unitAmount, 0) . ' ' : '') . $unitName;
-    }
-    
     /**
      * converts eg. months with only one digit with leading zero
      * @param int $number
